Add comment functionality and handle errors
