+++ backport/embed.go
@@ -1,0 +1,6 @@
+package main
+
+import _ "embed"
+
+//go:embed comment.tmpl
+var commentTemplate string
